+++ e2e/e2e-app/src/testConfig.ts
@@ -5,5 +5,4 @@ export type PackageName =
   | 'e2e-react'
   | 'e2e-solid'
   | 'e2e-svelte'
-  | 'e2e-vue2'
   | 'e2e-vue3';