+++ packages/core/src/helpers/component-file-extensions.ts
@@ -99,8 +99,6 @@ export const getComponentFileExtensionForTarget = (args: Args): string => {
     case 'swift':
       return '.swift';
     case 'vue':
-    case 'vue2':
-    case 'vue3':
       return '.vue';
     case 'webcomponent':
       return '.ts';