+++ packages/core/src/parsers/builder/builder.ts
@@ -4,7 +4,7 @@ import { BuilderContent, BuilderElement } from '@builder.io/sdk';
 import json5 from 'json5';
 import { mapKeys, merge, omit, omitBy, sortBy, upperFirst } from 'lodash';
 import traverse from 'traverse';
-import { MitosisComponent, MitosisState, hashCodeAsString } from '../..';
+import { MitosisComponent, MitosisState, blockToMitosis, hashCodeAsString } from '../..';
 import { Size, sizeNames, sizes } from '../../constants/media-sizes';
 import { createSingleBinding } from '../../helpers/bindings';
 import { capitalize } from '../../helpers/capitalize';
@@ -483,6 +483,7 @@ type BuilderToMitosisOptions = {
   context?: { [key: string]: any };
   includeBuilderExtras?: boolean;
   preserveTextBlocks?: boolean;
+  includeSpecialBindings?: boolean;
 };
 
 export const builderElementToMitosisNode = (
@@ -490,6 +491,8 @@ export const builderElementToMitosisNode = (
   options: BuilderToMitosisOptions,
   _internalOptions: InternalOptions = {},
 ): MitosisNode => {
+  const { includeSpecialBindings = true } = options;
+
   if (block.component?.name === 'Core:Fragment') {
     block.component.name = 'Fragment';
   }
@@ -576,7 +579,7 @@ export const builderElementToMitosisNode = (
     return mapper(block, options);
   }
 
-  const bindings: any = {};
+  const bindings: MitosisNode['bindings'] = {};
   const children: MitosisNode[] = [];
 
   if (blockBindings) {
@@ -586,9 +589,9 @@ export const builderElementToMitosisNode = (
       }
       const useKey = key.replace(/^(component\.)?options\./, '');
       if (!useKey.includes('.')) {
-        bindings[useKey] = {
+        bindings[useKey] = createSingleBinding({
           code: (blockBindings[key] as any).code || blockBindings[key],
-        };
+        });
       } else if (useKey.includes('style') && useKey.includes('.')) {
         const styleProperty = useKey.split('.')[1];
         // TODO: add me in
@@ -620,7 +623,21 @@ export const builderElementToMitosisNode = (
       const value = block.component.options[key];
       const valueIsArrayOfBuilderElements = Array.isArray(value) && value.every(isBuilderElement);
 
-      if (typeof value === 'string') {
+      const transformBldrElementToBinding = (item: BuilderElement) => {
+        const node = builderElementToMitosisNode(item, {
+          ...options,
+          includeSpecialBindings: false,
+        });
+
+        // For now, stringify to Mitosis nodes even though that only really works in React, due to syntax overlap.
+        // the correct long term solution is to hold on to the Mitosis Node, and have a plugin for each framework
+        // which processes any Mitosis nodes set into the attribute and moves them as slots when relevant (Svelte/Vue)
+        return blockToMitosis(node, {}, null as any);
+      };
+
+      if (isBuilderElement(value)) {
+        bindings[key] = createSingleBinding({ code: transformBldrElementToBinding(value) });
+      } else if (typeof value === 'string') {
         properties[key] = value;
       } else if (valueIsArrayOfBuilderElements) {
         const childrenElements = value
@@ -630,18 +647,14 @@ export const builderElementToMitosisNode = (
             }
             return true;
           })
-          .map((item) => builderElementToMitosisNode(item, options));
-        children.push({
-          '@type': '@builder.io/mitosis/node',
-          name: 'Slot',
-          meta: {},
-          scope: {},
-          bindings: {},
-          properties: { name: key },
-          children: childrenElements,
-        });
+          .map(transformBldrElementToBinding);
+
+        const strVal =
+          childrenElements.length === 1 ? childrenElements[0] : `<>${childrenElements.join('')}</>`;
+
+        bindings[key] = createSingleBinding({ code: strVal });
       } else {
-        bindings[key] = { code: json5.stringify(value) };
+        bindings[key] = createSingleBinding({ code: json5.stringify(value) });
       }
     }
   }
@@ -657,7 +670,7 @@ export const builderElementToMitosisNode = (
     if (binding.startsWith('component.options') || binding.startsWith('options')) {
       const value = blockBindings[binding];
       const useKey = binding.replace(/^(component\.options\.|options\.)/, '');
-      bindings[useKey] = { code: value };
+      bindings[useKey] = createSingleBinding({ code: value });
     }
   }
 
@@ -667,7 +680,7 @@ export const builderElementToMitosisNode = (
       block.tagName ||
       ((block as any).linkUrl ? 'a' : 'div'),
     properties: {
-      ...(block.component && { $tagName: block.tagName }),
+      ...(block.component && includeSpecialBindings && { $tagName: block.tagName }),
       ...(block.class && { class: block.class }),
       ...properties,
     },
@@ -675,11 +688,11 @@ export const builderElementToMitosisNode = (
       ...bindings,
       ...actionBindings,
       ...(styleString && {
-        style: { code: styleString },
+        style: createSingleBinding({ code: styleString }),
       }),
       ...(css &&
         Object.keys(css).length && {
-          css: { code: JSON.stringify(css) },
+          css: createSingleBinding({ code: JSON.stringify(css) }),
         }),
     },
   });