+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @builder.io/mitosis-cli
 
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies [9d6e019]
+- Updated dependencies [9d6e019]
+  - @builder.io/mitosis@0.1.1
+
+## 0.1.0
+
+### Minor Changes
+
+- 5dfd7cd: Breaking Change: remove vue2 and vue3 generators, keeping only the default vue generator (which generates vue3).
+
+### Patch Changes
+
+- Updated dependencies [5dfd7cd]
+  - @builder.io/mitosis@0.1.0
+
+## 0.0.147
+
+### Patch Changes
+
+- Updated dependencies [4e49454]
+  - @builder.io/mitosis@0.0.147
+
 ## 0.0.146
 
 ### Patch Changes